requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15.0,<3.16
xgboost>=2.1.4,<2.2
pandas>=2.2.0,<2.3
psycopg2>=2.9,<2.10
drf-spectacular>=0.27.0,<0.28
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
openai>=1.0.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0